Update master merge with dev updates (#363)
