+++ documents/CMakeLists.txt
@@ -8,11 +8,7 @@ set(DOXYGEN_INPUT_LIST ${CMAKE_SOURCE_DIR}/documents/DeveloperGuide
                        ${CMAKE_SOURCE_DIR}/source/MaterialXGenGlsl
                        ${CMAKE_SOURCE_DIR}/source/MaterialXGenGlsl/Nodes
                        ${CMAKE_SOURCE_DIR}/source/MaterialXGenOsl
-                       ${CMAKE_SOURCE_DIR}/source/MaterialXGenOsl
-                       ${CMAKE_SOURCE_DIR}/source/MaterialXGenArnold
                        ${CMAKE_SOURCE_DIR}/source/MaterialXRender
-                       ${CMAKE_SOURCE_DIR}/source/MaterialXRender/Handlers
-                       ${CMAKE_SOURCE_DIR}/source/MaterialXRender/ShaderValidators
                        ${CMAKE_SOURCE_DIR}/source/MaterialXRenderHw
                        ${CMAKE_SOURCE_DIR}/source/MaterialXRenderGlsl
                        ${CMAKE_SOURCE_DIR}/source/MaterialXRenderOsl
